+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.5.1] - 2023-05-23
+
+- Show link count on non-admin accounts
+- Add new link to top of list
+- Make checkbox more visible in dark mode
+
 ## [1.5.0] - 2023-05-18
 
 - Add backup functionality
@@ -109,3 +115,4 @@ First release, production ready
 - Main framework of Lynx
 - Import/Export
 - Add ShareX support
+